+++ modules/canvaskit/WasmCommon.h
@@ -36,6 +36,7 @@ using Float32Array = emscripten::val;
 // This doesn't stop us from using these as different types; e.g. a float* can be treated as an
 // SkPoint* in some APIs.
 using WASMPointerF32 = uintptr_t;
+using WASMPointerI32 = uintptr_t;
 using WASMPointerU8  = uintptr_t;
 using WASMPointerU16 = uintptr_t;
 using WASMPointerU32 = uintptr_t;