+++ modules/canvaskit/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - `Canvas.getDeviceClipBounds` (skbug.com/13347)
+
 ## [0.34.0] - 2022-05-05
 
 ### Breaking